model running in train.py

+++ utils/trainer.py
@@ -162,8 +162,8 @@ class NeuralNetworkClassifier:
                 for x, y in loader["train"]:
                     b_size = y.shape[0]
                     total += y.shape[0]
-                    x = x.to(self.device) if isinstance(x, torch.Tensor) else [i.to(self.device) for i in x]
-                    y = y.to(self.device)
+                    x = x.type(torch.LongTensor).to(self.device) if isinstance(x, torch.Tensor) else [i.to(self.device) for i in x]
+                    y = y.type(torch.LongTensor).to(self.device)
 
                     pbar.set_description(
                         "\033[36m" + "Training" + "\033[0m" + " - Epochs: {:03d}/{:03d}".format(epoch+1, epochs)
@@ -190,8 +190,8 @@ class NeuralNetworkClassifier:
                         self.model.eval()
                         for x_val, y_val in loader["val"]:
                             val_total += y_val.shape[0]
-                            x_val = x_val.to(self.device) if isinstance(x_val, torch.Tensor) else [i_val.to(self.device) for i_val in x_val]
-                            y_val = y_val.to(self.device)
+                            x_val = x_val.type(torch.LongTensor).to(self.device) if isinstance(x_val, torch.Tensor) else [i_val.to(self.device) for i_val in x_val]
+                            y_val = y_val.type(torch.LongTensor).to(self.device)
 
                             val_output = self.model(x_val)
                             val_loss = self.criterion(val_output, y_val)
@@ -235,8 +235,8 @@ class NeuralNetworkClassifier:
                 for x, y in (loader):
                     b_size = y.shape[0]
                     total += b_size
-                    x = x.to(self.device) if isinstance(x, torch.Tensor) else [i.to(self.device) for i in x]
-                    y = y.to(self.device)
+                    x = x.type(torch.LongTensor).to(self.device) if isinstance(x, torch.Tensor) else [i.to(self.device) for i in x]
+                    y = y.type(torch.LongTensor).to(self.device)
 
                     pbar.set_description("\033[32m"+"Evaluating"+"\033[0m")
                     pbar.update(b_size)
